[subsminder] rename utils -> subs_minder folder
