--- commands/subs-minder/.gitignore
@@ -1,2 +1,0 @@
-data.json
-raw_response.json
