+++ commands/subs-minder/lib/subs_minder/.env.rb
