+++ commands/subs-minder/lib/subs_minder/fetch.rb
