+++ commands/subs-minder/lib/subs_minder/fetch_to_file.rb
