+++ commands/subs-minder/lib/subs_minder/utils.rb
